fix: refactored Interim storage step's file path
