+++ frontend/src/components/SeedlotRegistrationSteps/InterimStep/definitions.ts
